endpoint: add common header fields to struct buffer
We ended up adding a lot of fields in the stack in order to route
messages and even parsing the message again on logendpoint just because
we need more fields to take some decisions.

Instead of doing that, start filling out new fields in struct buf that
have the info about the message the data field points to. Once other
parts of the code are adapted we can stop passing those fields.